Add min-width to copy text
