+++ src/components/widgets/Clipboard/Clipboard.less
@@ -14,6 +14,7 @@
     font-weight: 600;
     border: none;
     border-radius: 0.25rem;
+    padding: 0 0.5rem;
 
     @backgroundColor: #6ab889;
     background: rgba(
@@ -26,6 +27,11 @@
     > svg {
       margin-left: 0.25rem;
     }
+  }
+
+  .Clipboard__children {
+    min-width: 5rem;
+    text-align: right;
   }
 
   &:hover .Clipboard__children {
