+++ src/components/widgets/Clipboard/util.js
@@ -1,0 +1,14 @@
+export const copyContentsToClipboard = node => {
+    node.contentEditable = true;
+    const range = document.createRange();
+    range.selectNodeContents(node);
+
+    const selection = window.getSelection();
+    selection.removeAllRanges();
+    selection.addRange(range);
+
+    const didCopy = document.execCommand('copy');
+    document.getSelection().removeAllRanges();
+    node.contentEditable = false;
+    return didCopy;
+};
